move a lot of files around
