+++ Procfile
@@ -1,1 +1,3 @@
-web: npm start
+web: yarn start
+worker: yarn start:worker
+scheduler: yarn start:scheduler
