+++ package.json
@@ -15,6 +15,8 @@
     "watch:client": "webpack --watch",
     "build": "webpack",
     "start": "node src/index.js",
+    "start:worker": "node src/workers/index.js",
+    "start:scheduler": "node src/scheduler.js",
     "test": "mocha --watch --recursive \"./src/**/tests/*.js\""
   },
   "engines": {
