+++ src/logger.js
@@ -1,7 +1,5 @@
 const winston = require('winston');
 
-const IS_PRODUCTION = process.env.NODE_ENV === 'production';
-
 const logger = winston.createLogger({
   transports: [
     new winston.transports.Console(),
